Fix JPEG writer filename analyzer violation
